Implement ReportRepo.git_clone()
Allows a ReportRepo to be created from a remote Git repository (i.e.,
clone from GitHub).

Adds GitPython as a dependency.

Note: automodapi is having trouble with ReportRepo now that it has a
git_clone class method. I'm removing the warning-as-error build setting
to continue moving forward, but I think this is an upstream Sphinx
extension issue.