Create nbreport register command
This command registers a new report repo with LSST the Docs so that
report instances can be published.

This uses the POST /nbreport/reports/ endpoint of
sqre-uservice-nbreport.

To do this, I introduced two new keys to the metadata config file:

- git_repo (url of the git repo)
- git_repo_subpath (optional, path of the report repo within the git
  repo)

git_repo is needed to create the report's Product on LSST the Docs.

Information about the product in LSST the Docs is persisted back into
the repo's nbreport.yaml file:

- ltd_product,
- published_url,
- ltd_url