Add context to nbreport.yaml, and that to nb meta
When a template is rendered, the Jinja template context gets added to
nbreport.yaml under the 'context' key.

The content of nbreport.yaml is also added to the metadata of the
notebook under the 'nbreport' key.

This will allow the server to access notebook metadata when it
publishes the report.